+++ packages/lumx-core/src/js/types/index.ts
@@ -6,6 +6,7 @@ export type { HasCloseMode } from './HasCloseMode';
 export type { HasTheme } from './HasTheme';
 export type { GenericProps } from './GenericProps';
 export type { HeadingElement } from './HeadingElement';
+export type { KebabCase } from './KebabCase';
 export type { Point } from './Point';
 export type { Predicate } from './Predicate';
 export type { RectSize } from './RectSize';