+++ packages/lumx-core/src/js/utils/classNames/bem/index.ts
@@ -1,0 +1,2 @@
+export { block } from './block';
+export { element } from './element';