+++ packages/lumx-core/src/js/utils/classNames/index.ts
@@ -3,3 +3,4 @@ export * from './color';
 export * from './typography';
 export * from './spacing';
 export * from './visually-hidden';
+export * as bem from './bem';